Update catalog controller with Californica terms
This PR changes the catalog controller to use the
californica terms. The `solrizer` gem was added
because it is used to determine the names in californica.

The blacklight locale was also updated to use the correct
labels. There may be additional work around the locale/labels
that needs to be done.