Add a footer with the version
This adds a responsive footer to the page
that had the standard DCE version displayed
at the bottom. This also changes the header
to include an SVG logo, which is the
first part of adding a logo, but it will need
a PNG fallback.

Connected to UCLALibrary/ursus#10
Connected to UCLALibrary/amalgamated-samvera#132