Change app name to Distinctive Digital Collections.
This changes the blacklight locale and adds another
property to the locale. The application name is
UCLA Library Distinctive Digital Collections and the
`short_application_name` is Distinctive Digital Collections.
To use those in views, do this:

`<%= t('blacklight.application_name') %>`

or

`<%= t('blacklight.short_application_name') %>`

This replaces the second part of the logo with
text using CSS and the proper web font to replicate
what was there before. This is because the name is going
to change quite a bit.

Connected to UCLALibrary/ursus#99