Render creative commons licenses with links and logo
Remove the license field from the regular metadata display
of an object.
Instead, render the license field with its own specific helper.
This helper will detect a creative commons license and
render it with the CC logo, link, and language.
This is an extensible solution that can be added to as more
collections are added and more licenses are used. For now,
only the CC-BY 4.0 license is implemented.