Fix citation modal
There seems to be an incompatibility with the `boostrap`
gem used by blacklight and the bootstrap JS that is
required for making modal windows in Blacklight.
This adds the current bootstrap 3 js to the public folder and loads it
in a way that restores the modal functionality.

Connected to UCLALibrary/ursus#187