Use the californica db and solr instance in docker env
This adds an additional `docker-compose-with-californica.yml`
to use the same network as californica. After this is
merged, you can have californica running with `docker-compose`
up and then start ursus with this new config. This is so that they can
share the same solr instance and be running at the same
time.