Use iframe for loading UV on page
Loading UV directly in Blacklight is causing
conflicts becuase it includes its own versions
of jQuery UI, Bootrap JS, etc. Calling UV
in its iframe mode will isolate it and keep it
from interfering with the JS on the page. That
will be important when upgrading blacklight!

This PR removes the blob of /uv code from
/public and uses the `package.json` file to
install uv from the node package.

The travis build is changed so that it runs
`yarn install` before running tests to move
uv to the correct folder.

Before this PR is merged we should make sure that
`yarn` is installed on the production builds and
is installed in the ursus docker image.