Add yarn to Dockerfile
This adds the nodesource and yarn repos to apt
in the Dockerfile and installs them.

`npm` is the default node package manager, but rails
5.1+ is designed to work with `yarn`:

https://edgeguides.rubyonrails.org/5_1_release_notes.html#yarn-support

`yarn` is faster, and DCE has experience running it in production.

We'll need to have an ansible role that installs yarn on the ursus
production machines before deploying any commits that use
`yarn` to install dependencies.

Connected to #236