Restore UV config
This restores the UV config that was UV
before we switched to using the UV `iframe`.

The configuration for the view is in `config/uv`.

If you want to make changes to those files, they will
be used after running `yarn install`.

Connected to UCLALibrary/amalgamated-samvera#286
Connected to UCLALibrary/ursus#278