Restore fullscreen button
This commit makes the fullscreen button visible
on the work page.

There is a feature spec that checks for this and
the absence of the download button.

chrome-driver was converting the `<br>`'s that
are also tested in the view_work spec into whitespace
so I moved those tests into a seperate file that doesn't
use chrome-driver.

There is a `manifest.json` in public that is used
in the testing environment so that we can get UV to load
correctly.

This also updates the .travis.yml to include chromium and
chromedriver and the Dockerfile as well.

Connected to UCLALibrary/ursus#278