Remove previous font
This removes the Proxima Nova font that
was being used before the redesign and
replaces it with the Helvetica/Arial
font-stack used in the C5 recommendations.

If you need to change the font, it should
be done in the `_variables` file where it
overrides the correct Bootstrap variable.

Related to UCLALibrary/ursus#307