Thumbnail Fixes
This changes the way thumbnail URLs are
generated to work with Blacklight 7 and the new
version of `blacklight-gallery`.

The blacklight list view uses the new presenter way
of getting the URL and `blacklight-gallery` uses
the previous helper methods. This overrides the
presenter methods so that the correct URL will be
used and changes the presenter so that it uses
the correct `alt` tag value.

Connected to #308