Presenters for secondary metadata terms
This adds a presenter that creates a new
document hash that contains only the secondary
metadata terms.

This is controlled with a yml file in the `config`
directory. This is so that if the terms that
show up on the side need to change they can
be configured easily.

There is also a presenter that does the reverse --
it removes the secondary terms from the document
so that you can get a hash of only the primary
terms.

Connected to UCLALibrary/ursus#331