Add pa11y-ci to travis
This runs `pa11y-ci` on travis on the homepage.

Right now the error threshold is set at 10 errors.

There are 9 errors on the homepage now, so if more
are introduced the build will break. We'll eventually
want to reduce that threshold and add more pages
to the test.

Connected to URS-215