Use human_readable_resource_type instead of resource_type
There were changes made to Californica to the resource_type so we have to make the changes in Ursus too so the display of the resouce types in the facets and the view will be correct.

Once records are imported and indexed with https://github.com/UCLALibrary/californica/pull/594,

'resource_type` will store URI values.

English labels will be stored in `human_readable_resource_type` instead.

---

+ modified:   .rubocop.yml
+ modified:   app/controllers/catalog_controller.rb
+ modified:   config/materials_metadata.yml
+ modified:   spec/controllers/catalog_controller_spec.rb
+ modified:   spec/features/facet_labels_spec.rb
+ modified:   spec/features/search_results_spec.rb
+ modified:   spec/features/view_work_spec.rb