Re-size UV dynamically
This adds a script that adjusts the
width of the UV display to the width
of the container.

This needs to be done in JS because
UV needs a fixed width/height to load
on the page.

This will allow larger screens to view
more of the image and for smaller screens
to view the image without horizontal scrolling.

Connected to URS-365