Fix the date sort
This changes the date sorting and display so that
the user is able to sort by the creation date
of an item both ascending and descending. This
also changes the display field for the date on the
index page to display the date that was used
for sorting.

This also adds the ARK field to the default
search so that users can search by the ARK.

Connected to URS-406
Connected to URS-409