Change fielded search config
This changes the fielded search to the simplest possbile
config. Previously, it used a more specific config that
was working in the specs and in the development enviornment,
but did not work when deployed to production servers.

This change was tested on `ursus-dev` and is working across
the environments.

Connected to URS-400