Add tabs to collection block
This adds tabs with metadata, and other
content to the collection page.

Possible next steps:

Changing the default page to the gallery
view may result in funk behavor. Is it
always being the gallery view even if you selected
list view before a refresh OK?

Some of the metadata blocks say Item Overview and About
the Item. Should that be changed to collection

Connected to URS-404