remove stdin_open, tty: true from docker-compose.yml
These settings were added to support running byebug inside the container. I am removing them because they are not the best way to accomplish this, and we are running into trouble with docker on Travis. This might help.

The recommended way to access byebug is to shut down the web container and start it with `docker-compose up`:

```
docker-compose stop web
docker-compose run --service-ports web
```