Remove "Cite this item" from item pages until issues can be resolved
Connected to [URS-511](https://jira.library.ucla.edu/browse/URS-511)

Currently, the "Cite this item" has several issues that are not acceptable for the Dec. 10 launch. We would like the feature removed until we can resolve the issues.

Acceptance

"Cite this item" feature is not available on item pages in Ursus.

**I just commented it out so it will be there when we want to add it back in and fix it.**

---

Changes to be committed:
+ modified: `app/views/catalog/_document_action.html.erb`
+ modified: `app/views/catalog/_uv.html.erb`
+ modified: `spec/system/next_previous_on_show_spec.rb`
+ modified: `spec/system/view_work_spec.rb`