remove collection banner
The collection banner partial is currently being loaded (though it doesn't render anything). In the process, it makes a call to Californica (for branding info) and raises an error if this fails.

This commit disables the collection banner, but leaves related code (which won't be invoked). At some point we shoudl clean that up.