Edit text on About page
+ Our Services
+ UCLA Library Criteria for Digital Projects

Both of these sections are really outdated and not necessary to have published on the site at this time. It's possible in the future we will want to publish revised versions but not in 2019. In the meantime, it's better to be silent on these topics.

---

Changes to be committed:
modified:   app/views/static/about.html.erb