Don't check member_ids when forward canvas index
When forwarding a canvas ID from an item show page URL parameter to Universal Viewer, https://github.com/UCLALibrary/ursus/pull/698 checked to make sure that number was less than the total number of children in solr / hyrax.

We are currently importing many Manuscripts into californica without their ChildWork / page objects, since what matters is that they are in the manifest and we can import them to hyrax later. Thus, the check will not be accurate and in fact will result in a rails error when member_ids_ssim is nil.

Unfortunately, this will cause Universal Viewer to render a blank page if the `cv` parameter is invalid.