Use forward arks in URLs
This solution is kind of a hack. Changing SolrDocument.unique_key to 'ark_ssi' causes Blacklight to generate URLs using the ARK, but doesn't affect routing / document retrieval, which happens pretty deep in BL's internals. So for retrieval, I modified CatalogController's #show and #enforce_show_permissions methods to transform the ark from the URL parameter to a reversed-ark id.

The significant downside to this approach is that it now requires that ark_ssi be populated and that id be a reversed transformation of it.

I tried to get BL to retrieve documents via the 'ark_ssi' field instead of 'id' (the reversed ark), but couldn't figure it out. Someone with more solr knowledge might be able to but they would have to study the Blacklight::SearchService and Blacklight::Solr::Repository fields.