fix bug with CatalogController#cannonical_url_redirect
CatalogController#cannonical_url_redirect recognizes old-style URLs based on the hyrax id (`id` in solr) and redirects to a new-style URL using the ark.

Since hyrax also uses the Solr index to store permissions objects and other data, not everything in Solr is a digital library object with an ARK. A crawler, apparently trying random strings, used the hyrax id of one such object in the URL, raising an exception in the redirect code:
https://rollbar.com/uclalibrary/UrsusProd/items/129/

This fixes that exception by:
1. ensuring that `nil` can be handled in the `solr_document_path` helper function.
2. overriding SolrDocument.find to raise an error if it's superclass's .find method returns such an object without an ark.

CatalogController#cannonical_url_redirect is also cleaned up.