homepage update (#1144)
* homepage update

* fix: close anchor tag in homepage banner

* fix cashin collection link

* tinified images

* fix: delete ursus_collections.cy.js

This test ensured that the collection pages linked from the homepage "featured collections" grid load successfully. It isn't useful because the data used on our test Docker image is different from the data on the production server. The new collections we will be featuring are present in production but not on the test image. Adding them to the test image is a not insignificant amount of work, and would do nothing to ensure that they are present in production.